+++ realm/realm-library/src/main/java/io/realm/internal/SharedRealm.java
@@ -26,6 +26,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import javax.annotation.Nullable;
 
 import io.realm.RealmConfiguration;
+import io.realm.exceptions.RealmException;
 import io.realm.internal.android.AndroidCapabilities;
 import io.realm.internal.android.AndroidRealmNotifier;
 
@@ -166,6 +167,21 @@ public final class SharedRealm implements Closeable, NativeObject {
         void onSchemaChanged();
     }
 
+    /**
+     * Callback function to be called from JNI by Object Store when the partial sync results returned.
+     */
+    @Keep
+    public abstract static class PartialSyncCallback {
+        private final String className;
+
+        protected PartialSyncCallback(String className) {
+            this.className = className;
+        }
+
+        public abstract void onSuccess(Collection results);
+        public abstract void onError(RealmException error);
+    }
+
     private final OsRealmConfig osRealmConfig;
     private final long nativePtr;
     final NativeContext context;
@@ -220,7 +236,7 @@ public final class SharedRealm implements Closeable, NativeObject {
      */
     public static SharedRealm getInstance(OsRealmConfig.Builder configBuilder) {
         OsRealmConfig osRealmConfig = configBuilder.build();
-        ObjectServerFacade.getSyncFacadeIfPossible().wrapObjectStoreSessionIfRequired(osRealmConfig.getRealmConfiguration());
+        ObjectServerFacade.getSyncFacadeIfPossible().wrapObjectStoreSessionIfRequired(osRealmConfig);
 
         return new SharedRealm(osRealmConfig);
     }
@@ -346,6 +362,10 @@ public final class SharedRealm implements Closeable, NativeObject {
 
     public boolean isAutoRefresh() {
         return nativeIsAutoRefresh(nativePtr);
+    }
+
+    public void registerPartialSyncQuery(String query, PartialSyncCallback callback) {
+        nativeRegisterPartialSyncQuery(nativePtr, callback.className, query, callback);
     }
 
     public RealmConfiguration getConfiguration() {
@@ -476,6 +496,26 @@ public final class SharedRealm implements Closeable, NativeObject {
         callback.onInit(new SharedRealm(nativeSharedRealmPtr, osRealmConfig));
     }
 
+    /**
+     * Called from JNI when the partial sync callback is invoked from the ObjectStore.
+     * @param error if the partial sync query failed to register.
+     * @param nativeResultsPtr pointer to the {@code Results} of the partial sync query.
+     * @param callback the callback registered from the user to notify the success/error of the partial sync query.
+     */
+    @SuppressWarnings("unused")
+    private void runPartialSyncRegistrationCallback(@Nullable String error, long nativeResultsPtr,
+                                                           PartialSyncCallback callback) {
+        if (error != null) {
+            callback.onError(new RealmException(error));
+        } else {
+            @SuppressWarnings("ConstantConditions")
+            Table table = getTable(Table.getTableNameForClass(callback.className));
+            Collection results = new Collection(this, table, nativeResultsPtr, true);
+            callback.onSuccess(results);
+        }
+    }
+
+
     private static native void nativeInit(String temporaryDirectoryPath);
 
     private static native long nativeGetSharedRealm(long nativeConfigPtr, RealmNotifier notifier);
@@ -536,4 +576,7 @@ public final class SharedRealm implements Closeable, NativeObject {
     private static native long nativeGetSchemaInfo(long nativePtr);
 
     private static native void nativeRegisterSchemaChangedCallback(long nativePtr, SchemaChangedCallback callback);
+
+    private native void nativeRegisterPartialSyncQuery(
+            long nativeSharedRealmPtr, String className, String query, PartialSyncCallback callback);
 }