+++ realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
@@ -89,6 +89,7 @@ public class SyncConfiguration extends RealmConfiguration {
     private final String serverCertificateFilePath;
     private final boolean waitForInitialData;
     private final OsRealmConfig.SyncSessionStopPolicy sessionStopPolicy;
+    private final boolean isPartial;
 
     private SyncConfiguration(File directory,
                                 String filename,
@@ -118,7 +119,8 @@ public class SyncConfiguration extends RealmConfiguration {
                                 @Nullable
                                 String serverCertificateFilePath,
                                 boolean waitForInitialData,
-                                OsRealmConfig.SyncSessionStopPolicy sessionStopPolicy
+                                OsRealmConfig.SyncSessionStopPolicy sessionStopPolicy,
+                                boolean isPartial
     ) {
         super(directory,
                 filename,
@@ -146,6 +148,7 @@ public class SyncConfiguration extends RealmConfiguration {
         this.serverCertificateFilePath = serverCertificateFilePath;
         this.waitForInitialData = waitForInitialData;
         this.sessionStopPolicy = sessionStopPolicy;
+        this.isPartial = isPartial;
     }
 
     /**
@@ -360,6 +363,21 @@ public class SyncConfiguration extends RealmConfiguration {
     }
 
     /**
+     * Whether this configuration is for a partial synchronization Realm.
+     * Partial synchronization allows a synchronized Realm to be opened in such a way that
+     * only objects requested by the user are synchronized to the device. You can use it by setting
+     * the {@link Builder#partialRealm()}, opening the Realm, and then calling
+     * {@link Realm#subscribeToObjects(Class, String, Realm.PartialSyncCallback)} with the type of
+     * object you're interested in, a string containing a query determining which objects you want
+     * to subscribe to, and a callback which will report the results.
+     *
+     * @return {@code true} to open a partial synchronization Realm {@code false} otherwise.
+     */
+    public boolean isPartialRealm() {
+        return isPartial;
+    }
+
+    /**
      * Builder used to construct instances of a SyncConfiguration in a fluent manner.
      */
     public static final class Builder  {
@@ -394,6 +412,7 @@ public class SyncConfiguration extends RealmConfiguration {
         @Nullable
         private String serverCertificateFilePath;
         private OsRealmConfig.SyncSessionStopPolicy sessionStopPolicy = OsRealmConfig.SyncSessionStopPolicy.AFTER_CHANGES_UPLOADED;
+        private boolean isPartial = false;
         /**
          * Creates an instance of the Builder for the SyncConfiguration.
          * <p>
@@ -819,6 +838,15 @@ public class SyncConfiguration extends RealmConfiguration {
             return this;
         }
 
+        /**
+         * Setting this will open a partially synchronized Realm.
+         * @see #isPartialRealm()
+         */
+        public SyncConfiguration.Builder partialRealm() {
+            this.isPartial = true;
+            return this;
+        }
+
         private String MD5(String in) {
             try {
                 MessageDigest digest = MessageDigest.getInstance("MD5");
@@ -964,7 +992,8 @@ public class SyncConfiguration extends RealmConfiguration {
                     serverCertificateAssetName,
                     serverCertificateFilePath,
                     waitForServerChanges,
-                    sessionStopPolicy
+                    sessionStopPolicy,
+                    isPartial
             );
         }